Break out the `install` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`install` command.
